Oh my god.
I don't even want to talk about it. Do not ask me about accelerationTracker.cs in my life. Don't.

For the actual important changes, I reorganized the file structure and also added a test script that allows the developer to move the ball with WASD and Space. It includes a global Input System as well.