Created a Worldspace UI for objects.
Included is a yield instruction to track when the mouse click is lifted. Currently almost perfect, only bug is that it kind of rotates. I might have to just disable the Y axis constraint and make a update function always update it above the ball. I don't want to though since that's very performance heavy, I feel like it's definitely doable just with a parent-like structure.