Fixed acceleration Vector bug
Just in case acceleration ever has miniscule amounts of movement that messes with rotation.

In all honesty, a larger fix might need to be at play, as the mini values can still occur. Maybe we need to round vector axis values to a certain decimal to avoid this entirely. For now, this is a good loophole.