Refactored Structure & Finished Acceleration Realm
It's unfortunate that both changes are combined, but there should be ZERO problems that can come about. If there are any and we need to revert this commit, I'll handle it.

In short, refactored the project structure to be more organized, and added to the Acceleration Realm experiment to make it better.